+++ CHANGELOG.md
@@ -6,11 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.4] - 2025-9-24
 ### Fixed
 - Track update time of segments in an MS3TraceList consistently
 
 ### Changed
 - Update MS3TraceList.pack() to use new libmseed function for performance
+  when packing with a specified `flush_idle_seconds` set.
 
 ## [0.0.3] - 2025-9-14
 ### Added
@@ -55,7 +58,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MS3Record class for individual records
 - CFFI-based bindings to libmseed
 
-[Unreleased]: https://github.com/EarthScope/pymseed/compare/v0.0.3...HEAD
+[Unreleased]: https://github.com/EarthScope/pymseed/compare/v0.0.4...HEAD
+[0.0.4]: https://github.com/EarthScope/pymseed/releases/tag/v0.0.4
 [0.0.3]: https://github.com/EarthScope/pymseed/releases/tag/v0.0.3
 [0.0.2]: https://github.com/EarthScope/pymseed/releases/tag/v0.0.2
 [0.0.1]: https://github.com/EarthScope/pymseed/releases/tag/v0.0.1